+++ Source/FreeImage/freeimage_misc.cpp
@@ -6,7 +6,7 @@
 void FreeImage_utf8_to_utf16(std::wstring &out, const char *in)
 {
     out.resize(strlen(in));
-    int newLen = MultiByteToWideChar(CP_UTF8, 0, in, out.size(), &out[0], out.size());
+    int newLen = MultiByteToWideChar(CP_UTF8, 0, in, (int)out.size(), &out[0], (int)out.size());
     out.resize(newLen);
 }
